+++ ceygen/tests/__init__.py
@@ -10,3 +10,4 @@ from ceygen.tests.test_core import *
 from ceygen.tests.test_dispatch import *
 from ceygen.tests.test_elemwise import *
 from ceygen.tests.test_lu import *
+from ceygen.tests.test_reductions import *